requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
black==23.7.0
blinker==1.6.2
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
ConfigArgParse==1.5.5
coverage==7.2.7
cryptography==41.0.2
exceptiongroup==1.1.2
fastapi==0.100.0
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==23.7.0
geventhttpclient==2.0.9
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.15.1
Mako==1.2.4
MarkupSafe==2.1.3
msgpack==1.0.5
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
pluggy==1.2.0
psutil==5.9.5
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==2.0.3
pydantic-settings==2.0.2
pydantic_core==2.3.0
PyJWT==2.3.0
pytest==7.4.0
python-creditcard @ git+https://github.com/maistodos/python-creditcard.git@6b6bcc5d975985337c35e5c1a4f87d861bcac731
python-dotenv==1.0.0
pyzmq==25.1.0
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
Werkzeug==2.3.6
zope.event==5.0
zope.interface==6.0
